+++ .github/workflows/notebook_tests.yml
@@ -1,0 +1,28 @@
+name: NotebookTests
+
+on: [push]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 4
+      matrix:
+        python-version: [3.7]
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -U pip wheel;
+        pip install -r requirements.txt;
+        pip install treon;
+        pip install --no-deps .
+    - name: Run Notebooks
+      run: |
+        treon
